+++ src/bos/operators/power_off_graceful.py
@@ -2,7 +2,7 @@
 #
 # MIT License
 #
-# (C) Copyright 2021-2023 Hewlett Packard Enterprise Development LP
+# (C) Copyright 2021-2024 Hewlett Packard Enterprise Development LP
 #
 # Permission is hereby granted, free of charge, to any person obtaining a
 # copy of this software and associated documentation files (the "Software"),
@@ -53,8 +53,9 @@ class GracefulPowerOffOperator(BaseOperator):
         ]
 
     def _act(self, components):
-        component_ids = [component['id'] for component in components]
-        pcs.soft_off(component_ids)
+        if components:
+            component_ids = [component['id'] for component in components]
+            pcs.soft_off(component_ids)
         return components