+++ src/bos/server/controllers/v2/components.py
@@ -24,7 +24,7 @@
 import connexion
 import logging
 
-from bos.common.utils import get_current_timestamp
+from bos.common.utils import exc_type_msg, get_current_timestamp
 from bos.common.tenant_utils import get_tenant_from_header, get_tenant_component_set, tenant_error_handler
 from bos.common.values import Phase, Action, Status, EMPTY_STAGED_STATE, EMPTY_BOOT_ARTIFACTS
 from bos.server import redis_db_utils as dbutils
@@ -54,6 +54,7 @@ def get_v2_components(ids="", enabled=None, session=None, staged_session=None, p
         try:
             id_list = ids.split(',')
         except Exception as err:
+            LOGGER.error("Error parsing component IDs: %s", exc_type_msg(err))
             return connexion.problem(
                 status=400, title="Error parsing the ids provided.",
                 detail=str(err))
@@ -179,7 +180,7 @@ def put_v2_components():
         ComponentArray.from_dict(data)  # noqa: E501
     except Exception as err:
         msg="Provided data does not follow API spec"
-        LOGGER.exception(msg)
+        LOGGER.error("%s: %s", msg, exc_type_msg(err))
         return connexion.problem(status=400, title=msg,detail=str(err))
 
     components = []
@@ -220,7 +221,7 @@ def patch_v2_components():
             ComponentArray.from_dict(data)  # noqa: E501
         except Exception as err:
             msg="Provided data does not follow API spec"
-            LOGGER.exception(msg)
+            LOGGER.error("%s: %s", msg, exc_type_msg(err))
             return connexion.problem(status=400, title=msg,detail=str(err))
         return patch_v2_components_list(data)
     elif type(data) == dict:
@@ -230,13 +231,14 @@ def patch_v2_components():
             ComponentsUpdate.from_dict(data)  # noqa: E501
         except Exception as err:
             msg="Provided data does not follow API spec"
-            LOGGER.exception(msg)
+            LOGGER.error("%s: %s", msg, exc_type_msg(err))
             return connexion.problem(status=400, title=msg,detail=str(err))
         return patch_v2_components_dict(data)
 
+    LOGGER.error("Unexpected data type %s", str(type(data)))
     return connexion.problem(
-        status=400, title="Error parsing the data provided.",
-        detail="Unexpected data type {}".format(str(type(data))))
+       status=400, title="Error parsing the data provided.",
+       detail="Unexpected data type {}".format(str(type(data))))
 
 
 def patch_v2_components_list(data):
@@ -246,11 +248,13 @@ def patch_v2_components_list(data):
         for component_data in data:
             component_id = component_data['id']
             if component_id not in DB or not _is_valid_tenant_component(component_id):
+                LOGGER.warning("Component %s could not be found", component_id)
                 return connexion.problem(
                     status=404, title="Component not found.",
                     detail="Component {} could not be found".format(component_id))
             components.append((component_id, component_data))
     except Exception as err:
+        LOGGER.error("Error loading component data: %s", exc_type_msg(err))
         return connexion.problem(
             status=400, title="Error parsing the data provided.",
             detail=str(err))
@@ -268,6 +272,7 @@ def patch_v2_components_dict(data):
     ids = filters.get("ids", None)
     session = filters.get("session", None)
     if ids and session:
+        LOGGER.warning("Multiple filters provided")
         return connexion.problem(
             status=400, title="Only one filter may be provided.",
             detail="Only one filter may be provided.")
@@ -275,6 +280,7 @@ def patch_v2_components_dict(data):
         try:
             id_list = ids.split(',')
         except Exception as err:
+            LOGGER.error("Error parsing the IDs provided: %s", exc_type_msg(err))
             return connexion.problem(
                 status=400, title="Error parsing the ids provided.",
                 detail=str(err))
@@ -289,6 +295,7 @@ def patch_v2_components_dict(data):
         id_list = [component["id"] for component in get_v2_components_data(session=session, tenant=get_tenant_from_header())]
         LOGGER.debug("patch_v2_components_dict: %d IDs found for specified session", len(id_list))
     else:
+        LOGGER.warning("No filter provided")
         return connexion.problem(
             status=400, title="Exactly one filter must be provided.",
             detail="Exactly one filter may be provided.")
@@ -308,6 +315,7 @@ def get_v2_component(component_id):
     """Used by the GET /components/{component_id} API operation"""
     LOGGER.debug("GET /v2/components/%s invoked get_v2_component", component_id)
     if component_id not in DB or not _is_valid_tenant_component(component_id):
+        LOGGER.warning("Component %s could not be found", component_id)
         return connexion.problem(
             status=404, title="Component not found.",
             detail="Component {} could not be found".format(component_id))
@@ -337,7 +345,7 @@ def put_v2_component(component_id):
         Component.from_dict(data)  # noqa: E501
     except Exception as err:
         msg="Provided data does not follow API spec"
-        LOGGER.exception(msg)
+        LOGGER.error("%s: %s", msg, exc_type_msg(err))
         return connexion.problem(status=400, title=msg,detail=str(err))
     data['id'] = component_id
     data = _set_auto_fields(data)
@@ -365,14 +373,16 @@ def patch_v2_component(component_id):
         Component.from_dict(data)  # noqa: E501
     except Exception as err:
         msg="Provided data does not follow API spec"
-        LOGGER.exception(msg)
+        LOGGER.error("%s: %s", msg, exc_type_msg(err))
         return connexion.problem(status=400, title=msg,detail=str(err))
 
     if component_id not in DB or not _is_valid_tenant_component(component_id):
+        LOGGER.warning("Component %s could not be found", component_id)
         return connexion.problem(
             status=404, title="Component not found.",
             detail="Component {} could not be found".format(component_id))
     if "actual_state" in data and not validate_actual_state_change_is_allowed(component_id):
+        LOGGER.warning("Not able to update actual state")
         return connexion.problem(
             status=409, title="Actual state can not be updated.",
             detail="BOS is currently changing the state of the node,"
@@ -407,6 +417,7 @@ def delete_v2_component(component_id):
     """Used by the DELETE /components/{component_id} API operation"""
     LOGGER.debug("DELETE /v2/components/%s invoked delete_v2_component", component_id)
     if component_id not in DB or not _is_valid_tenant_component(component_id):
+        LOGGER.warning("Component %s could not be found", component_id)
         return connexion.problem(
             status=404, title="Component not found.",
             detail="Component {} could not be found".format(component_id))
@@ -433,10 +444,11 @@ def post_v2_apply_staged():
                     response["succeeded"].append(xname)
                 else:
                     response["ignored"].append(xname)
-            except Exception as err:
-                LOGGER.error(f"An error was encountered while attempting to apply stage for node {xname}: {err}")
+            except Exception:
+                LOGGER.exception("An error was encountered while attempting to apply stage for node %s", xname)
                 response["failed"].append(xname)
     except Exception as err:
+        LOGGER.error("Error parsing request data: %s", exc_type_msg(err))
         return connexion.problem(
             status=400, title="Error parsing the data provided.",
             detail=str(err))